gofmt -tabs=false -tabwidth=2
I am not very happy with the formatting gofmt provides, but as there's
the cult of gophers who seems to love this tool, it makes sense to try
to avoid "why not gofmt!?" discussions and not distract existing go
users with differing code style.

Next step is obviously to write logstash-forwarder in !Go just to
avoid formatting wars, right? ;)