Use net.JoinHostPort() for stringifying an ip:host tuple.
Apart from reducing the amount of code this change actually fixes
a cosmetic bug since the Go library might represent IPv4 addresses
with 16 bytes instead of 4. If that happens to be the case the IPv4
address would previously be surrounded by square brackets as if it
were an IPv6 address.