Only perform DNS lookup for hostnames.
When the native Go DNS resolver is used (e.g. when cross-compiling),
passing a dotted-quad IP address to net.LookupHost() fails. Such
calls were always made for the configured destination hosts so
for affected environments it became impossible to configure
logstash-forwarder to connect to IP addresses.

We now call net.LookupHost() only if net.ParseIP() fails, indicating
that the string actually is a hostname that needs to be resolved.
This closes #371.